Minimized conversion wait times
If the last command issued on the 1-Wire is a "convert temperature", then reads can be done until the value returned is 1, which means that the conversion is finished. The conversion times in the manuals are maximum values, actual conversion times are much faster and the only reason to delay the maximum times are when one or more devices are parasitically powered.